Painstakingly got the PLL to work with a 12MHz crystal
This is closer than the low IRC could get me. I found that with the IRC
my slowest speed was reported as 24.08MHz on my scope and 24.13MHz with
the highest speed. This seemed a little off so I installed a crystal
(with no load capacitors...probably not a good thing) and now I'm at
24.02MHz which I think might be good enough.